Moved authenticate_user and validate_and_revoke_refresh_token to validator.py file
